Added sound effect to atk1. fixed the audo problems and the audio conversion files. Audio should be easy to add now, just need to find sounds we want to use, then we convert them and then call play_sfx or play_bgm.

+++ src/audio_assets/attack_1_sword_swing_1.h
@@ -9,6 +9,6 @@
 
 #define ATTACK_1_SWORD_SWING_1_LENGTH 9024
 
-const short attack_1_sword_swing_1[ATTACK_1_SWORD_SWING_1_LENGTH];
+const int attack_1_sword_swing_1[ATTACK_1_SWORD_SWING_1_LENGTH];
 
 #endif /* ATTACK_1_SWORD_SWING_1_H */
